pyproject.toml
[tool.poetry]
name = "fpg-dashboard-resultados"
version = "0.1.0"
description = ""
authors = ["olv-leo <olv.leo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.43.1"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
plotly = "^6.0.0"
folium = "^0.19.5"
streamlit-folium = "^0.24.0"
psycopg2-binary = "^2.9.10"
cryptography = "^44.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
